requirements.txt
python-telegram-bot
git+https://github.com/kssb-cc/kssb-menu
